Implemented day 3
Note that only part 1 has been implemented. The reasoning is on the file
header. Moreover, we are not printing 24-bit results yet because that
would entail modifying the somewhat terrifying vendor code on
`vendor/bcd16.s`. Since this will be a recurring theme, I guess that I
will have to come up with a solution sooner or later.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>